fix: Fixed a bug where phecodes with less than the minimum case count were still being run
Discovered a bug in the if statement within the run_logit_regression function on line 153. Whenever there were extra individuals in the counts file that were not in the covariate file, the phecode_obj.case_count used for comparision would be over inflated. We should have been comparing the case count in the analysis cohort which can be done after the creation of the column phecode_status in the covariate dataframe.