fix(logit-maximum-iteration): Allowed user to set the logit maximum iteration
Discovered that most of the phecodes that we failing to converge were the result of the maximum iteration threshold being exceeded in the logit function. Added a flag that allows the user to set the maximum iteration count. The default value is now 200 which seems to allow most phecodes to convery with little detriment to runtime.