refactor(run_PheWAS.py): - jtb324 - 12/9/25 - 7:54pm --- On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   src/pyphewas/parser/cmd_parser.py
	modified:   src/pyphewas/run_PheWAS.py
	modified:   tests/integration_tests.py

changed default value of phecode-version flag to be the string "None":
---
The previous version of the code required you to pass the
phecode-version as either "phecodeX" or "phecode1.2" even if you weren't
going to use either of the PheCode mappings. Now the user can set it to
None. If the user sets the value to none and also doesn't provide any
phecode description file then the program will just return N/As for the
phecode description. This change makes the behavior of the argument
clearer