fix(run_PheWAS.py): - jtb324 - 12/12/25 - 7:26pm --- On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   pdm.lock
	modified:   pyproject.toml
	modified:   src/pyphewas/run_PheWAS.py

fixed type mismatch for regression that caused pickling bug:
---
The run_regression function was passing a polars dataframe to the
regression models. Statsmodels doesn't support polars they only support
pandas dataframes. This error was raising a NotImplemented exception,
which was then not able to be pickled by the multiprocessing library
which was raising a PicklingError. This error is now fixed by converting
the covariates polars dataframe to the covariates_df pandas dataframe.