feature(run_PheWAS.py): - jtb324 - 12/16/25 - 4:46pm --- On branch firth Changes to be committed: 	modified:   pdm.lock 	modified:   pyproject.toml 	modified:   src/pyphewas/run_PheWAS.py
added method for running firth regression when perfect separation is
encountered:
---
Sometimes because of the correlation of the phecodes the logistic model
will run into the perfect separation error. One solution to this probelm
is firth regression which now has been implemented in the firthmodels
package. This package uses the scipy api and not statsmodel so we can
use the patsy library to generate the appropriate input structure