feature(run_PheWAS.py): jtb324 - 12/18/25 - 11:24am --- On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   pyproject.toml
	modified:   src/pyphewas/parser/cmd_parser.py
	modified:   src/pyphewas/run_PheWAS.py

added the ability to log state when the program encounters errors:
---
For debugging purposes I added the ability for the user to log the
state of the program when an error is encountered in the regression.
Currently it creates a tmp directory in the current directory and write
information about the regression to the file and then it saves the
covariate matrix. This argument makes it easier to reproduce what the
program is running and to compare with similar functionality in R