feature(run_PheWAS.py): - jtb324 - 12/18/25 - 8:37am --- On branch firth Changes to be committed: 	new file:   src/pyphewas/model_formatters.py 	modified:   src/pyphewas/run_PheWAS.py
created function to format firth results and connect firth regression to
the rest of the program:
---
Finished connecting the firth regression to the rest of the script by
creating another function to format the results that is analagous to the
format_results function. This function accesses elements of the
regression model differently since it uses the scipy api instead of
statsmodel. Also added a column to the output indicating if firth
regression was used in the run