fix(run_PheWAS.py): - jtb324 - 12/19/25 - 10:22pm --- On branch firth Changes to be committed: 	modified:   pyproject.toml 	modified:   src/pyphewas/run_PheWAS.py 	modified:   tests/integration_tests.py
fixed a bug where firth was being used for linear & logistic regression:
---
The code wasn't checking to see if the model being used originally was
logistic before running the FirthLogisticModel. Firth is not made for
continuous variables so this was wrong behavior. There is a check now
that makes sure the regression is logistic before running firth