feature(run_PheWAS.py): - jtb324 - 12/20/25 - 2:45pm ---
On branch firth
Changes to be committed:
	modified:   README.md
	modified:   src/pyphewas/run_PheWAS.py

set the maximum number of child processes per worker:
---
In order to reduce the memory consumption using in the linear regression
model, I set the maxtasksperchild parameter in the linear regression
model to be 50 so that the thread will be restarted after so many jobs.
It seems the memory was caused by a memory leak in the underlying c
processes. Now the code uses between 10-15 GB of RAM per cpu.