fix(plotting.py): - jtb324 - 12/20/25 - 4:39pm --- On branch plot-fix Changes to be committed: 	modified:   src/pyphewas/plotting.py
fixed manhattan plot so infinite values were shown correctly:
---
There was a bug where the infinite values were not being replaced and
therefore they were just being dropped from the manhattan plot. I fixed
this and now the values are replaced with 1.02 times the max
non-infinite value in the data.